refactor test result building
